Encapsulate projectmanager
