+++ mindbender/tools/projectmanager/style.py
