--- mindbender/widgets/deselectabletreeview.py
@@ -1,16 +1,0 @@
-from ..vendor.Qt import QtWidgets, QtCore
-
-
-class DeselectableTreeView(QtWidgets.QTreeView):
-    """A tree view that deselects on clicking on an empty area in the view"""
-
-    def mousePressEvent(self, event):
-
-        index = self.indexAt(event.pos())
-        if not index.isValid():
-            # clear the selection
-            self.clearSelection()
-            # clear the current index
-            self.setCurrentIndex(QtCore.QModelIndex())
-
-        QtWidgets.QTreeView.mousePressEvent(self, event)
